Tcp Custom Op and Torch->Tcp conversions + lit tests (#17)
Op definition for `tcp.custom_op` and lowerings for `aten.gather` and
`aten.index.Tensor_hacked_twin`.

This is useful to represent and lower custom/opaque operations that have
tensor inputs and outputs, but
1) lack a clear operational semantics, or 
2) have an operational semantic that isn't defined by Tcp (such as user
defined custom operations), or
3) need to "flow through" the middleend because they target pre-written
kernels available in the backend

As a follow-on, I'll look into making `torch -> tcp.custom_op`
conversions automatic as much as possible, to minimize the overhead in
adding individual conversion patterns manually. Then this could be the
last step in a frontend lowering pipeline, to have all remaining
`torch.aten` ops converted.