Add op for Gather (#68)
This PR adds the `tcp.gather` op. The design and semantics of this op
are defined in
[docs/gather.md](https://github.com/cruise-automation/mlir-tcp/blob/main/docs/gather.md).

More specifically, this PR includes:
* The definition of the `tcp.gather` op.
* Lowering from Torch to Tcp for the `torch.gather` op.
* Lowering from Tcp to Linalg for the `tcp.gather` op.
* Lit tests for the op and the lowerings.
* AOT test the Tcp gather op.
* Renames the op in the design document.

This PR does not update the lowering we have from aten.gather to
tcp.custom("aten.gather") pattern
[here](https://github.com/cruise-automation/mlir-tcp/blob/b812717001121df6974687e2b2e7e82b900390dd/lib/Conversion/TorchToTcp/TcpCustomOp.cpp#L29-L45).
We should be able to control how `aten.gather` gets lowered by fixing
the order in which the conversion passes are called in the pipeline.