Handle `tcp.bind_symbolic_shape` ops in fusion algorithm (#82)
The current fusion algorithm does not handle `tcp.bind_symbolic_shape`
ops very well because these ops cause an op to have multiple downstream
uses. This change handles these ops "specially". We collect all the uses
of an op which only specify the shape of the output into a separate use
category. We then move the bind shape ops along with the original op if
it was moved to another region.

---------

Co-authored-by: Srinath Avadhanula <srinath.avadhanula@getcruise.com>