Slight improvement to fusion (#86)
As discussed, a small improvement to the fusion algorithm to account for
the case when we have multiple uses of a definition but all those uses
already belong to the group we are in. This allows us to fuse "diamond
patterns" of multiple uses into a single group. See the update to the
lit test for the improvement.

Note that this still is not a "maximal" fusion which can create groups
with multiple returns etc.

---------

Co-authored-by: Srinath Avadhanula <srinath.avadhanula@getcruise.com>