Add support for customizing const inlining and ops with regions to IsolateGroupOps pass (#87)
A few small improvements to the `IsolateGroupOps` pass:
- Provide a hook to customize whether a const like op used in the
`tcp.group` will get copied into the group or whether it will be passed
in as an input argument. The main pass does always returns true so this
is a non-functional change in `mlir-tcp`.
- The previous version did not handle ops with contained regions and
block arguments (such as an `scf.forall` inside the `tcp.group`). We now
handle this (see updated lit test).

---------

Co-authored-by: Srinath Avadhanula <srinath.avadhanula@getcruise.com>